add: bunch of stuff

+++ src/student/domain/student.py
@@ -11,10 +11,10 @@ class Student(BaseModel):
         new_execution = SessionExecution(seq_num=len(self.session_executions)+1)
         self.session_executions.append(new_execution)
 
-    def record_macro_feedback(self, feedback: float) -> None:
+    def record_macro_feedback(self, feedback: int) -> None:
         # Recording is always done to the last executing session
         self.session_executions[-1].record_macro_feedback(feedback)
 
-    def record_micro_feedback(self, feedback: float) -> None:
+    def record_micro_feedback(self, feedback: int) -> None:
         # Recording is always done to the last executing session
         self.session_executions[-1].record_micro_feedback(feedback)
